Restructure file tree to prevent parent projects' TS compiler from compiling library sources
