--- index.ts
@@ -1,15 +1,0 @@
-require('intro.js/introjs.css');
-
-require('./styles/ontodia.css');
-require('./src/svgui/svgui.css');
-
-export * from './src/ontodia/data/model';
-export * from './src/ontodia/data/provider';
-export * from './src/ontodia/data/demo/provider';
-export * from './src/ontodia/data/sparql/provider';
-export * from './src/ontodia/data/sparql/graphBuilder';
-export * from './src/ontodia/diagram/elements';
-export * from './src/ontodia/diagram/elementViews';
-export * from './src/ontodia/diagram/model';
-export * from './src/ontodia/diagram/view';
-export { Workspace, Props as WorkspaceProps } from './src/ontodia/workspace/workspace';
