+++ src/examples/sparqlConstruct.ts
@@ -1,7 +1,7 @@
 import { createElement, ClassAttributes } from 'react';
 import * as ReactDOM from 'react-dom';
 
-import { Workspace, WorkspaceProps, SparqlDataProvider, GraphBuilder } from '../../index';
+import { Workspace, WorkspaceProps, SparqlDataProvider, GraphBuilder } from '../index';
 
 require('jointjs/css/layout.css');
 require('jointjs/css/themes/default.css');
